refactor(playground): split block tree
