+++ packages/agent-playground/src/components/playground/block-visualization/block-tree/index.tsx
@@ -1,0 +1,2 @@
+export { BlockTree } from './block-tree';
+export type { IBlockTreeProps } from './types';
